Documentation of existing QC process to date for GTPSS and a number of small fixes in support

+++ src/cnodc/programs/nodb_intake/qcworker.py
@@ -25,6 +25,8 @@ class NODBQCWorker(QueueWorker):
             'max_batch_size': None,
             'max_buffer_size': None,
             'target_buffer_size': None,
+            'input_is_source_file': False,
+            'use_station_batching': False,
             'next_queue': None,
             'review_queue': None,
         })
@@ -43,6 +45,7 @@ class NODBQCWorker(QueueWorker):
                 'review_queue': self.get_config('review_queue')
             },
             use_source_file=self.get_config('input_is_source_file', False),
+            use_station_batcher=self.get_config('use_station_batching', False),
             processor_uuid=self.process_uuid,
         )
 
@@ -60,6 +63,7 @@ class QCProcessor(PayloadProcessor):
                  submitter_kwargs: dict,
                  processor_uuid: str,
                  use_source_file: bool = False,
+                 use_station_batcher: bool = False,
                  **kwargs):
         self._test_suite: BaseTestSuite = dynamic_object(test_suite)(**test_suite_kwargs, test_runner_id=processor_uuid)
         super().__init__(
@@ -69,6 +73,8 @@ class QCProcessor(PayloadProcessor):
             processor_uuid=processor_uuid,
             **kwargs
         )
+        # Always use the station batcher when processing a source file
+        self._use_station_batcher = use_station_batcher or use_source_file
         self._use_source_file = use_source_file
         self._batcher_kwargs = batcher_kwargs or {}
         self._submitter_kwargs = submitter_kwargs or {}
@@ -76,12 +82,17 @@ class QCProcessor(PayloadProcessor):
     def _process(self):
         self._test_suite.set_db_instance(self._db)
         submitter = NODBBatchSubmitter(self._db, self._current_payload, **self._submitter_kwargs)
-        separator = ResultBatcher(submitter, **self._batcher_kwargs)
+        separator = (
+            StationResultBatcher(batch_submitter=submitter, **self._batcher_kwargs)
+            if self._use_station_batcher else
+            SimpleResultBatcher(batch_submitter=submitter, **self._batcher_kwargs)
+        )
         if not self._use_source_file:
             batch = self.load_batch_from_payload()
             self._process_records(batch.stream_working_records(self._db, lock_type=LockType.FOR_NO_KEY_UPDATE), separator)
             separator.flush_all()
-            self._db.delete_object(batch)
+            if separator.remove_original_batch:
+                self._db.delete_object(batch)
             self._current_item.mark_complete(self._db)
             self._db.commit()
         else:
@@ -91,12 +102,14 @@ class QCProcessor(PayloadProcessor):
                 separator
             )
             separator.flush_all()
+            self._current_item.mark_complete(self._db)
+            self._db.commit()
 
     def _process_records(self, records: t.Iterable[structures.NODBWorkingRecord], separator):
         for wr, dr, outcome, is_modified in self._test_suite.process_batch(records):
             if is_modified:
                 self._update_working_record(wr, dr)
-            separator.add_result(wr.working_uuid, dr, outcome)
+            separator.add_result(wr, dr, outcome)
 
     def _update_working_record(self,
                                working_record: structures.NODBWorkingRecord,
@@ -132,23 +145,16 @@ class NODBBatchSubmitter:
                  last_payload: WorkflowPayload,
                  next_queue: str = None,
                  review_queue: str = None,
-                 failure_queue: str = None,
                  next_step_queue: str = 'nodb_continue'):
         self._db = db
         self._next_step = next_step_queue
         self._previous_payload = last_payload
         self._next_queue = next_queue
-        self._failure_queue = failure_queue
         self._review_queue = review_queue
 
-    def submit_batch(self, working_uuids: list[str], batch_outcome: BatchOutcome, group_key: t.Optional[str] = None):
-        batch = structures.NODBBatch()
-        batch.batch_uuid = str(uuid.uuid4())
-        batch.status = structures.BatchStatus.QUEUED
-        self._db.insert_object(batch)
-        structures.NODBWorkingRecord.bulk_set_batch_uuid(self._db, working_uuids, batch.batch_uuid)
+    def submit_existing_batch(self, batch_id: str, batch_outcome: BatchOutcome, group_key: t.Optional[str] = None):
         queue_name = self._get_queue_name(batch_outcome)
-        payload = BatchPayload(batch.batch_uuid)
+        payload = BatchPayload(batch_id)
         self._previous_payload.update_for_propagation(payload)
         if queue_name is not None:
             self._db.create_queue_item(
@@ -157,7 +163,6 @@ class NODBBatchSubmitter:
                 data=payload.to_map(),
                 unique_item_key=group_key
             )
-            self._db.commit()
         else:
             if group_key is not None:
                 payload.headers['forward-unique-item-key'] = group_key
@@ -167,7 +172,15 @@ class NODBBatchSubmitter:
                 queue_name=self._next_step,
                 data=payload.to_map(),
             )
-            self._db.commit()
+
+    def submit_batch(self, working_uuids: list[str], batch_outcome: BatchOutcome, group_key: t.Optional[str] = None):
+        batch = structures.NODBBatch()
+        batch.batch_uuid = str(uuid.uuid4())
+        batch.status = structures.BatchStatus.QUEUED
+        self._db.insert_object(batch)
+        structures.NODBWorkingRecord.bulk_set_batch_uuid(self._db, working_uuids, batch.batch_uuid)
+        self.submit_existing_batch(batch.batch_uuid, batch_outcome, group_key)
+        self._db.commit()
 
     def _get_subqueue_name(self) -> t.Optional[str]:
         if 'manual-subqueue' in self._previous_payload.headers:
@@ -183,39 +196,22 @@ class NODBBatchSubmitter:
             raise ValueError('Invalid batch outcome type')
 
 
-class ResultBatcher:
+class BaseResultBatcher:
 
     def __init__(self,
                  batch_submitter: NODBBatchSubmitter,
                  max_batch_size: int = None,
                  max_buffer_size: int = None,
-                 target_buffer_size: int = None):
+                 target_buffer_size: int = None,
+                 remove_original_batch: bool = False):
+        self.remove_original_batch = not remove_original_batch
         self._submitter = batch_submitter
-        self._result_batches = {}
-        self._current_total = 0
-        self._max_batch_size = max_batch_size if max_batch_size is not None and max_batch_size > 0 else 100
-        self._max_total_size = max_buffer_size if max_buffer_size is not None and max_buffer_size > 0 else 1000000
-        self._target_total_size = target_buffer_size if target_buffer_size is not None and target_buffer_size > 0 else 250000
+        self._max_batch_size = max_batch_size if max_batch_size is not None and max_batch_size > 0 else None
+        self._max_total_size = max_buffer_size if max_buffer_size is not None and max_buffer_size > 0 else None
+        self._target_total_size = target_buffer_size if target_buffer_size is not None and target_buffer_size > 0 else self._max_total_size
+        self.remove_original_batch = True
 
-    def add_result(self, working_uuid: str, record: ocproc2.DataRecord, outcome: ocproc2.QCResult):
-        group_key = self._generate_unique_group(record)
-        target_queue = self._target_queue(outcome)
-        batch_key = self._generate_batch_key(group_key)
-        if batch_key not in self._result_batches:
-            self._result_batches[batch_key] = [[], group_key, target_queue, 0]
-        elif target_queue == BatchOutcome.REVIEW_QUEUE:
-            self._result_batches[batch_key][2] = target_queue
-        self._result_batches[batch_key][0].append(working_uuid)
-        self._result_batches[batch_key][3] += 1
-        self._current_total += 1
-        if self._result_batches[batch_key][3] >= self._max_batch_size:
-            self.flush(batch_key)
-        self._check_auto_flush()
-
-    def _generate_batch_key(self, group_key: t.Optional[str]):
-        return hashlib.md5(group_key.encode('utf-8', 'replace')).hexdigest()
-
-    def _target_queue(self, outcome: ocproc2.QCResult) -> BatchOutcome:
+    def _determine_batch_outcome(self, outcome: ocproc2.QCResult) -> BatchOutcome:
         if outcome == ocproc2.QCResult.MANUAL_REVIEW:
             return BatchOutcome.REVIEW_QUEUE
         return BatchOutcome.NEXT_QUEUE
@@ -229,8 +225,68 @@ class ResultBatcher:
             return record.metadata.best_value('CNODCStationString')
         return None
 
+    def add_result(self, working: structures.NODBWorkingRecord, record: ocproc2.DataRecord, outcome: ocproc2.QCResult):
+        raise NotImplementedError
+
+    def flush_all(self):
+        raise NotImplementedError
+
+
+class SimpleResultBatcher(BaseResultBatcher):
+
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs, remove_original_batch=False)
+        self._batch_ids = {}
+
+    def add_result(self, working: structures.NODBWorkingRecord, record: ocproc2.DataRecord, outcome: ocproc2.QCResult):
+        if working.qc_batch_id is None:
+            raise ValueError('missing batch id')
+        batch_outcome = self._determine_batch_outcome(outcome)
+        if working.qc_batch_id not in self._batch_ids:
+            self._batch_ids[working.qc_batch_id] = [batch_outcome, self._generate_unique_group(record)]
+        elif batch_outcome == BatchOutcome.REVIEW_QUEUE:
+            self._batch_ids[working.qc_batch_id][0] = batch_outcome
+
+    def flush_all(self):
+        for existing_id in self._batch_ids:
+            self._submitter.submit_existing_batch(
+                batch_id=existing_id,
+                group_key=self._batch_ids[existing_id][1],
+                batch_outcome=self._batch_ids[existing_id][0]
+            )
+        self._batch_ids = {}
+
+
+class StationResultBatcher(BaseResultBatcher):
+
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs, remove_original_batch=True)
+        self._result_batches = {}
+        self._current_total = 0
+
+    def add_result(self, working: structures.NODBWorkingRecord, record: ocproc2.DataRecord, outcome: ocproc2.QCResult):
+        group_key = self._generate_unique_group(record)
+        target_queue = self._determine_batch_outcome(outcome)
+        batch_key = self._generate_batch_key(group_key)
+        if batch_key not in self._result_batches:
+            self._result_batches[batch_key] = [[], group_key, target_queue, 0]
+        elif target_queue == BatchOutcome.REVIEW_QUEUE:
+            self._result_batches[batch_key][2] = target_queue
+        self._result_batches[batch_key][0].append(working.working_uuid)
+        self._result_batches[batch_key][3] += 1
+        self._current_total += 1
+        if self._max_batch_size is not None and self._result_batches[batch_key][3] >= self._max_batch_size:
+            self.flush(batch_key)
+        self._check_auto_flush()
+
+    def _generate_batch_key(self, group_key: t.Optional[str]):
+        x = {
+            'group_key': group_key or '',
+        }
+        return hashlib.md5('\x1F'.join(f'{y}={x[y]}' for y in x).encode('utf-8', 'replace')).hexdigest()
+
     def _check_auto_flush(self):
-        if self._current_total >= self._max_total_size:
+        if self._max_total_size is not None and self._current_total >= self._max_total_size:
             for bn in self._flush_queue():
                 self.flush(bn)
                 if self._current_total < self._target_total_size:
